Make components separate
Make components separate to make pages able to be server rendered instead of client rendered
