+++ src/components/addRecipeForm/AddRecipeForm.module.css
@@ -1,0 +1,41 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+}
+
+.label,
+.button {
+  margin-top: 1rem;
+}
+
+.input {
+  width: 50%;
+  height: 1.6rem;
+  border-radius: var(--border-radius);
+  padding: 0.4rem;
+  margin-top: 0.2rem;
+  font-family: var(--font-mono);
+}
+
+.textarea {
+  font-family: var(--font-mono);
+  width: 100%;
+  height: 10rem;
+  margin-top: 0.2rem;
+  border-radius: var(--border-radius);
+  padding: 0.5rem;
+}
+
+.button {
+  width: 50%;
+  height: 2rem;
+  background-color: #4caf50;
+  color: white;
+  border: none;
+  cursor: pointer;
+  border-radius: var(--border-radius);
+  font-family: var(--font-mono);
+}
