+++ src/components/detailRecipe/DetailRecipe.module.css
@@ -1,7 +1,7 @@
 .recipeDiv {
-  display : flex;
-  flex-direction : column;
-  align-items : flex-start;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
   width: 100%;
   padding: 1rem;
   border-radius: var(--border-radius);
